refactor: DTO에 from 매서드 추가
